src/ml/ingest_data.py
import pandas as pd
import numpy as np
from datetime import datetime
import os
import yaml
import requests
import logging
from typing import Dict, List, Any, Optional
from abc import ABC, abstractmethod
from bs4 import BeautifulSoup

# Set up logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

class DataSource(ABC):
    """Abstract base class for data sources"""
    
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        self.source_name = config.get('source', 'Unknown')
        self.api_endpoint = config.get('api_endpoint', '')
        self.update_frequency = config.get('update_frequency', 'daily')
        
    @abstractmethod
    def fetch_data(self) -> pd.DataFrame:
        """Fetch data from the source"""
        pass
    
    @abstractmethod
    def validate_data(self, data: pd.DataFrame) -> bool:
        """Validate the fetched data"""
        pass
    
    def save_raw_data(self, data: pd.DataFrame, dataset_name: str):
        """Save raw data to appropriate directory"""
        timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
        filename = f"{dataset_name}_{timestamp}.parquet"
        source_type = self.__class__.__name__.lower().replace('source', '')
        filepath = os.path.join('data', 'raw', source_type, filename)
        os.makedirs(os.path.dirname(filepath), exist_ok=True)
        data.to_parquet(filepath, index=False)
        logger.info(f"Saved raw data to {filepath}")

class ABSSource(DataSource):
    """Australian Bureau of Statistics data source"""
    
    def fetch_data(self) -> pd.DataFrame:
        logger.info("Fetching ABS data...")
        
        # Fetch macro indicators
        macro_data = self._fetch_macro_indicators()
        
        # Fetch local area data
        local_data = self._fetch_local_data()
        
        # Combine the data
        combined_data = local_data.copy()
        for macro_metric in macro_data.columns:
            if macro_metric != 'postcode':
                # Add macro data as additional columns
                combined_data[f'macro_{macro_metric}'] = macro_data[macro_metric].iloc[0]
        
        return combined_data
    
    def _fetch_macro_indicators(self) -> pd.DataFrame:
        """Fetch national-level economic indicators"""
        url = 'https://www.abs.gov.au/statistics/economy/key-indicators'
        try:
            tables = pd.read_html(url)
            # Process and combine relevant tables
            macro_data = pd.DataFrame()
            
            # Extract relevant economic indicators
            for table in tables:
                if 'indicator' in table.columns.str.lower():
                    processed_table = self._process_macro_table(table)
                    macro_data = pd.concat([macro_data, processed_table], axis=1)
            
            return macro_data
        except Exception as e:
            logger.error(f"Error fetching macro data: {str(e)}")
            return pd.DataFrame()
    
    def _fetch_local_data(self) -> pd.DataFrame:
        """Fetch postcode-specific data"""
        local_data = []
        base_url = "https://www.abs.gov.au/census/find-census-data/quickstats/2021/POA"
        
        for postcode in ['2000', '2026', '2028']:  # Your target postcodes
            try:
                url = f"{base_url}{postcode}"
                response = requests.get(url)
                soup = BeautifulSoup(response.text, 'html.parser')
                
                # Extract local statistics
                data = {
                    'postcode': postcode,
                    'timestamp': datetime.now().strftime('%Y%m%d')
                }
                
                # Add extraction of specific metrics here
                # This will be customized based on the HTML structure
                
                local_data.append(data)
            except Exception as e:
                logger.error(f"Error fetching data for postcode {postcode}: {str(e)}")
        
        return pd.DataFrame(local_data)
    
    def _process_macro_table(self, table: pd.DataFrame) -> pd.DataFrame:
        """Process a macro indicator table"""
        # Clean column names
        table.columns = table.columns.str.strip().str.lower()
        
        # Convert numeric values
        for col in table.columns:
            if col != 'indicator':
                table[col] = pd.to_numeric(table[col].str.replace('%', '').str.replace(',', ''), errors='coerce')
        
        return table
    
    def validate_data(self, data: pd.DataFrame) -> bool:
        """Validate the fetched data"""
        required_columns = ['postcode', 'timestamp']
        return all(col in data.columns for col in required_columns)

class PropertySource(DataSource):
    """Domain API data source"""
    
    def fetch_data(self) -> pd.DataFrame:
        logger.info("Fetching property data...")
        # TODO: Implement Domain API integration
        # Mock data for now
        data = pd.DataFrame({
            'postcode': ['2000', '2026', '2028'],
            'median_price': [1200000, 2500000, 3500000],
            'growth_rate': [5.2, 4.8, 4.5],
            'total_listings': [150, 80, 45]
        })
        return data
    
    def validate_data(self, data: pd.DataFrame) -> bool:
        required_columns = ['postcode', 'median_price', 'growth_rate']
        return all(col in data.columns for col in required_columns)

class DataIngestion:
    """Main data ingestion coordinator"""
    
    def __init__(self):
        self.config = self._load_config()
        self.data_sources = self._initialize_sources()
        
    def _load_config(self) -> Dict[str, Any]:
        """Load configuration from YAML"""
        config_path = 'data/metadata/data_sources.yaml'
        with open(config_path, 'r') as f:
            return yaml.safe_load(f)
    
    def _initialize_sources(self) -> Dict[str, DataSource]:
        """Initialize all data sources"""
        return {
            'abs': ABSSource(self.config['abs_data']),
            'property': PropertySource(self.config['property_data'])
            # Add other sources as implemented
        }
    
    def ingest_all(self) -> Dict[str, pd.DataFrame]:
        """Ingest data from all sources"""
        results = {}
        for source_name, source in self.data_sources.items():
            try:
                data = source.fetch_data()
                if source.validate_data(data):
                    source.save_raw_data(data, source_name)
                    results[source_name] = data
                else:
                    logger.error(f"Data validation failed for {source_name}")
            except Exception as e:
                logger.error(f"Error ingesting data from {source_name}: {str(e)}")
        return results
    
    def process_data(self, raw_data: Dict[str, pd.DataFrame]) -> pd.DataFrame:
        """Process and combine data from all sources"""
        logger.info("Processing data...")
        
        # Start with property data as base
        if 'property' not in raw_data:
            raise ValueError("Property data is required but missing")
            
        combined_data = raw_data['property']
        
        # Merge with other sources
        for source_name, data in raw_data.items():
            if source_name != 'property':
                combined_data = combined_data.merge(
                    data,
                    on='postcode',
                    how='left'
                )
        
        # Calculate derived metrics
        combined_data['risk_score'] = self._calculate_risk_score(combined_data)
        
        # Save processed data
        timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
        output_path = os.path.join(
            'data',
            'processed',
            'features',
            f'processed_data_{timestamp}.parquet'
        )
        combined_data.to_parquet(output_path, index=False)
        logger.info(f"Processed data saved to {output_path}")
        
        return combined_data
    
    def _calculate_risk_score(self, data: pd.DataFrame) -> pd.Series:
        """Calculate risk score based on various metrics"""
        # Implement risk score calculation
        # This is a simplified version
        return (
            data['growth_rate'] * 0.3 +
            data['employment_rate'] * 0.2 +
            (data['total_listings'].rank(pct=True) * 100) * 0.1
        )

def main():
    ingestion = DataIngestion()
    raw_data = ingestion.ingest_all()
    processed_data = ingestion.process_data(raw_data)
    logger.info(f"Data ingestion complete! Processed {len(processed_data)} records")

if __name__ == "__main__":
    main() 