requirements.txt
pandas==2.2.3
scikit-learn==1.6.1
fastapi==0.115.11
uvicorn==0.34.0
xgboost==3.0.0
openai==1.66.3
python-dotenv==1.0.1
numpy==2.2.4
pydantic==1.10.13
joblib>=1.1.0
httpx>=0.24.0
starlette>=0.40.0,<0.47.0
beautifulsoup4==4.12.0
requests==2.31.0
PyYAML==6.0.1 